+++ src/components/Editor/Panes/StructurePane/Tree/PathList/Path.tsx
