+++ src/components/Editor/Panes/StructurePane/Tree/PathList/PathList.tsx
